Added pom.xml, separate enums classes, renamed Users to User
