+++ .gitignore
@@ -2,5 +2,6 @@
 /shelf/
 /workspace.xml
 /.idea/
+/target/
 /out/
 *.iml
