+++ src/main/java/com/company/Main.java
@@ -1,4 +1,4 @@
-package main.java.com.company;
+package com.company;
 
 public class Main {
