+++ src/main/java/com/company/model/user/enums/Rights.java
@@ -1,0 +1,6 @@
+package com.company.model.user.enums;
+
+public enum Rights {
+    USER,
+    ADMIN
+}
