+++ src/main/java/com/company/model/user/enums/Status.java
@@ -1,0 +1,6 @@
+package com.company.model.user.enums;
+
+public enum Status {
+    ACTIVE,
+    BLOCKED
+}
