+++ src/main/java/com/company/service/StoreManager.java
@@ -1,4 +1,4 @@
-package main.java.com.company.service;
+package com.company.service;
 
 public class StoreManager {
 }
