+++ src/main/java/com/company/service/UserOrder.java
@@ -1,4 +1,4 @@
-package main.java.com.company.service;
+package com.company.service;
 
 public class UserOrder {
 }
